server/requirements.txt
flask
python-dotenv
requests
chromadb
bs4
langchain_text_splitters
groq

